requirements.txt
fastapi
uvicorn
firebase-admin
opencv-python
mediapipe
numpy
pydantic
python-multipart
python-dotenv 